Add 'async' keyword to %%ipytest
